refactor comps
