+++ components/Swimlane.tsx
@@ -1,8 +1,9 @@
 /** @jsx h */
 
+import { h } from '../deps.ts';
 import { Page } from '../types/ApiPages.ts';
 import type { MappedAsset } from '../types/MappedAsset.ts';
-import { h } from '../deps.ts';
+import { ProgressBar } from './ProgressBar.tsx';
 
 export function Swimlane({s}: {s: Page['swimlanes'][0] & {items: any[]}}) {
   return (
@@ -30,21 +31,5 @@ function Item({item}: {item: MappedAsset}) {
     <h3>{item.title}</h3>
     <p class="hover">{item.subtitle}</p>
   </div>
-  );
-}
-
-function ProgressBar({start, end}: {start: Date | null, end: Date | null}) {
-  if (!start || !end) return '';
-  const now = Date.now();
-  const startTime = start.getTime();
-  const endTime = end.getTime();
-  if (now > endTime) return '';
-  if (now < startTime) return '';
-  const progress = Math.round((now - startTime) / (endTime - startTime) * 100);
-  const style = `height: 5px; width: ${progress}%; background: rgba(255 255 0 / 75%);`;
-  return (
-    <div style="border-radius: 50vw; overflow: hidden; background: rgba(100 100 100 / 50%); position: relative; bottom: 13px; width: 95%; margin: auto;">
-      <div style={style}></div>
-    </div>
   );
 }
